New envoy built image works

+++ packer/script.sh
@@ -20,21 +20,51 @@ sudo DEBIAN_FRONTEND=noninteractive apt-get --quiet -y install \
    ninja-build \
    nginx
 
+if ! ninja --version; then
+  echo 'Error could not find ninja builder'
+  exit 1
+fi
+
 sudo wget -qO /usr/local/bin/bazel https://github.com/bazelbuild/bazelisk/releases/download/v1.12.0/bazelisk-linux-amd64
 sudo chmod +x /usr/local/bin/bazel
 
+if ! bazel --version; then
+  echo 'Error could not find bazel builder'
+  exit 1
+fi
+
 sudo curl -s -LO https://go.dev/dl/go1.18.3.linux-amd64.tar.gz
-sudo rm -rf /usr/local/go
+if [ -d /usr/local/go ]; then
+   sudo rm -rf /usr/local/go
+fi
 sudo tar -C /usr/local -xzf go1.18.3.linux-amd64.tar.gz
 export PATH=$PATH:/usr/local/go/bin
-go version
+
+if ! go version; then
+  echo 'Error could not find go'
+  exit 1
+fi
 
 go install github.com/bazelbuild/buildtools/buildifier@5.1.0
 go install github.com/bazelbuild/buildtools/buildozer@5.1.0
 
+if [ -d ./envoy ]; then
+   sudo rm -rf ./envoy
+fi
+
 git clone -q --branch v1.21.0 https://github.com/envoyproxy/envoy.git
 cd envoy
 bazel build envoy
-sudo mv bazel-bin/source/exe/envoy-static /usr/local/bin/envoy
-envoy --version
+if [ -x bazel-bin/source/exe/envoy-static ]; then
+   sudo mv bazel-bin/source/exe/envoy-static /usr/local/bin/envoy
+fi
 cd ..
+sudo rm -rf ./envoy
+
+if ! envoy --version; then
+  echo 'Error could not find envoy'
+  exit 1
+fi
+
+echo 'Build successful'
+exit 0
